--- core/src/cg/group4/rewards/collectibles/collectibleSorters/package-info.java
@@ -1,4 +1,0 @@
-/**
- * 
- */
-package cg.group4.rewards.collectibles.collectibleSorters;
