+++ core/src/cg/group4/rewards/collectibles/collectible_sorters/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 
+ */
+package cg.group4.rewards.collectibles.collectible_sorters;
