+++ core/src/cg/group4/view/screen/CollectiblesScreen.java
@@ -1,14 +1,12 @@
 package cg.group4.view.screen;
 
-import java.util.ArrayList;
-
 import cg.group4.rewards.CollectibleDrawer;
+import cg.group4.rewards.Collection;
 import cg.group4.rewards.collectibles.Collectible;
-import cg.group4.rewards.collectibles.collectibleSorters.CollectibleSorter;
-import cg.group4.rewards.collectibles.collectibleSorters.SortByRarity;
+import cg.group4.rewards.collectibles.collectible_sorters.CollectibleSorter;
+import cg.group4.rewards.collectibles.collectible_sorters.SortByRarity;
 import cg.group4.view.screen_mechanics.ScreenLogic;
 import cg.group4.view.screen_mechanics.ScreenStore;
-
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.ScrollPane;
@@ -16,6 +14,8 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.ui.WidgetGroup;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
+
+import java.util.ArrayList;
 
 public class CollectiblesScreen extends ScreenLogic {
